sensor/requirements.txt
Flask
python-dotenv
psutil